# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Kaay entreprendre",
        "description": "Cet api va nous permettre de faire communiquer la partie front et la partie back de notre plateforme",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemes": {
          "BearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ], 
    "paths": {
...
        "/api/phase/{id}": {
            "get": {
                "tags": [
                    "Phases"
                ],
                "summary": "Récupérer une phase en particulier",
                "description": "Récupère la phase specifier avec l'id",
                "operationId": "getPhase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la phase à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des phases récupérée avec succès"
                    }
                }
            }
        },
...
        "/api/create_phase": {
            "post": {
                "tags": [
                    "Phases"
                ],
                "summary": "Ajouter une nouvelle phase",
                "description": "Ajoute une nouvelle phase avec les détails fournis",
                "operationId": "createphase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "guide_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "phase créé avec succès"
                    }
                }
            }
        },
        "/api/update_phase/{id}": {
            "post": {
                "tags": [
                    "Phases"
                ],
                "summary": "Mettre à jour une phase existant",
                "description": "Met à jour les détails d'une phase existant en fonction de l'ID fourni",
                "operationId": "updatephase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la phase à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "phase mise à jour avec succès"
                    }
                }
            }
        },
        "/api/archiver_phase/{id}": {
            "post": {
                "tags": [
                    "Phases"
                ],
                "summary": "Archiver une phase",
                "description": "Archive une phase en fonction de l'ID fourni",
                "operationId": "archivephase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du phase à archiver",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "phase archivé avec succès"
                    }
                }
            }
        },
...
        "/api/reponse/create": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Ajouter une nouvelle réponse",
                "description": "Ajoute une nouvelle réponse avec les détails fournis",
                "operationId": "createReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "commentaire_id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse ajoutée avec succès"
                    }
                }
            }
        },
        "/api/reponse/edit": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Mettre à jour une réponse existante",
                "description": "Met à jour les détails d'une réponse existante",
                "operationId": "updateReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse mise à jour avec succès"
                    }
                }
            }
        },
        "/api/reponse/archive": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Archiver une réponse",
                "description": "Archive une réponse en fonction de l'ID fourni",
                "operationId": "archiveReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse archivée avec succès"
                    }
                }
            }
        },
        "/api/reponse/delete": {
            "post": {
                "tags": [
                    "Réponses"
                ],
                "summary": "Supprimer une réponse",
                "description": "Supprime une réponse en fonction de l'ID fourni",
                "operationId": "destroyReponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "numeric"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse supprimée avec succès"
                    }
                }
            }
        },
...
    },
...
}